Define equal function globally
