+++ lib/hir-mir/src/lib.rs
@@ -31,7 +31,7 @@ pub use map_type_configuration::{
 };
 pub use string_type_configuration::StringTypeConfiguration;
 pub use test_module_configuration::TestModuleConfiguration;
-use transformation::{map_context, record_equal_function, record_hash_function};
+use transformation::{equal_operation, map_context, record_equal_function, record_hash_function};
 
 pub fn compile_main(
     module: &Module,
@@ -78,6 +78,7 @@ fn compile_module(
     let module = record_equal_function::transform(&context, &module)?;
     let module = record_hash_function::transform(&context, &module)?;
     let module = map_context::module::transform(&context, &module)?;
+    let module = equal_operation::module::transform(&context, &module)?;
 
     Ok((
         {
